requirements.txt
# python 3.8
absl-py==0.9.0
attrs==24.2.0
black==24.8.0
cachetools==5.5.0
certifi @ file:///croot/certifi_1725551672989/work/certifi
cfgv==3.4.0
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==1.3.0
cmake==3.25.0
deepdiff==8.0.1
distlib==0.3.9
docker-pycreds==0.4.0
enum34==1.1.10
filelock==3.13.1
flake8==5.0.4
flake8-bugbear==23.3.12
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
google-auth-oauthlib==1.0.0
grpcio==1.67.1
gym==0.17.2
identify==2.6.1
idna==3.4
importlib_metadata==8.5.0
Jinja2==3.1.3
lit==15.0.7
Markdown==3.7
MarkupSafe==2.1.5
mccabe==0.7.0
mock==5.1.0
mpmath==1.3.0
mpyq==0.2.5
mypy-extensions==1.0.0
networkx==3.0
nodeenv==1.9.1
numpy==1.21.0
oauthlib==3.2.2
orderly-set==5.2.2
packaging==24.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.3.6
portpicker==1.6.0
pre-commit==3.5.0
protobuf==3.20.0
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.9.1
pyflakes==2.5.0
pygame==2.6.1
pyglet==1.5.0
PySC2==3.0.0
PyYAML==6.0.2
requests==2.28.1
requests-oauthlib==2.0.0
rsa==4.9
s2clientprotocol==5.0.13.92440.3
s2protocol==5.0.13.92440.4
scipy==1.10.1
sentry-sdk==2.17.0
setproctitle==1.1.10
six==1.16.0
sk-video==1.1.10
-e git+https://github.com/oxwhirl/smacv2.git@577ab5a2cff2391f8df582da5731ea9cd6adf3c6#egg=SMACv2
smmap==5.0.1
sympy==1.13.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tomli==2.0.2
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
triton==2.0.0
typing_extensions==4.9.0
urllib3==1.26.13
virtualenv==20.27.1
wandb==0.18.6
websocket-client==1.8.0
Werkzeug==3.0.6
whichcraft==0.6.1
zipp==3.20.2
